Refactored to use router
